omit AutomationsWizard prefix in components
