+++ ui-v2/src/components/automations/automations-wizard/action-step/action-change-flow-run-state-fields/action-change-flow-run-state-fields.tsx
@@ -1,0 +1,11 @@
+import { ActionChangeFlowRunStateMessageField } from "./action-change-flow-run-state-message-field";
+import { ActionChangeFlowRunStateNameField } from "./action-change-flow-run-state-name-field";
+import { ActionChangeFlowRunStateStateField } from "./action-change-flow-run-state-state-field";
+
+export const ActionChangeFlowRunStateFields = () => (
+	<div>
+		<ActionChangeFlowRunStateStateField />
+		<ActionChangeFlowRunStateNameField />
+		<ActionChangeFlowRunStateMessageField />
+	</div>
+);
