+++ ui-v2/src/components/automations/automations-wizard/action-step/index.ts
@@ -1,0 +1,1 @@
+export { ActionStep } from "./action-step";
