--- ui-v2/src/components/automations/automations-wizard/automations-wizard-action-change-flow-run-state-fields/index.ts
@@ -1,3 +1,0 @@
-export { AutomationsWizardActionChangeFlowRunStateMessageField } from "./automations-wizard-action-change-flow-run-state-message-field";
-export { AutomationsWizardActionChangeFlowRunStateNameField } from "./automations-wizard-action-change-flow-run-state-name-field";
-export { AutomationsWizardActionChangeFlowRunStateStateField } from "./automations-wizard-action-change-flow-run-state-state-field";
