--- ui-v2/src/components/automations/automations-wizard/automations-wizard-action-step/automations-wizard-action-step.stories.tsx
@@ -1,14 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react";
-
-import { fn } from "@storybook/test";
-import { AutomationsWizardActionStep } from "./automations-wizard-action-step";
-
-const meta = {
-	title: "Components/Automations/Wizard/AutomationWizardActionStep",
-	component: AutomationsWizardActionStep,
-	args: { onSubmit: fn() },
-} satisfies Meta;
-
-export default meta;
-
-export const story: StoryObj = { name: "AutomationWizardActionStep" };
