--- ui-v2/src/components/automations/automations-wizard/automations-wizard-action-step/index.ts
@@ -1,1 +1,0 @@
-export { AutomationsWizardActionStep } from "./automations-wizard-action-step";
