--- ui-v2/src/components/automations/automations-wizard/automations-wizard-action-type-additional-fields.tsx
@@ -1,42 +1,0 @@
-import {
-	AutomationsWizardActionChangeFlowRunStateMessageField,
-	AutomationsWizardActionChangeFlowRunStateNameField,
-	AutomationsWizardActionChangeFlowRunStateStateField,
-} from "@/components/automations/automations-wizard/automations-wizard-action-change-flow-run-state-fields";
-
-import { useWatch } from "react-hook-form";
-import { ActionsSchema } from "./action-type-schemas";
-
-export const AutomationsWizardActionTypeAdditionalFields = () => {
-	const actionType = useWatch<ActionsSchema>({ name: "type" });
-	switch (actionType) {
-		case "change-flow-run-state":
-			return (
-				<>
-					<AutomationsWizardActionChangeFlowRunStateStateField />
-					<AutomationsWizardActionChangeFlowRunStateNameField />
-					<AutomationsWizardActionChangeFlowRunStateMessageField />
-				</>
-			);
-		case "run-deployment":
-		case "pause-deployment":
-		case "resume-deployment":
-			return <div>TODO Deployment</div>;
-		case "pause-work-queue":
-		case "resume-work-queue":
-			return <div>TODO Work Queue</div>;
-		case "pause-work-pool":
-		case "resume-work-pool":
-			return <div>TODO Work pool</div>;
-		case "pause-automation":
-		case "resume-automation":
-			return <div>TODO Automation</div>;
-		case "send-notification":
-			return <div>TODO send notification</div>;
-		case "cancel-flow-run":
-		case "suspend-flow-run":
-		case "resume-flow-run":
-		default:
-			return null;
-	}
-};
